1. move jscript_massage and js unit test javascript to ggadget/js 2. Add automake script for qt_script_runtime 3. Now Object in QtScriptRuntime can be called as function 4. GetCurrentFileAndLine in QtScriptRuntime works

+++ extensions/qt_script_runtime/tests/js_shell.cc
@@ -40,7 +40,7 @@ enum QuitCode {
 };
 QuitCode g_quit_code = DONT_QUIT;
 
-static const int kBufferSize = 655360;
+static const size_t kBufferSize = 655360;
 static char g_buffer[kBufferSize];
 static void Process(QScriptEngine *engine, const char *filename) {
   FILE *file;
